Refactor trace writer test modules
